% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_image_rscript.R
\name{crop_image_secondary}
\alias{crop_image_secondary}
\title{crop an image in R  [ for RGB or grey images ]}
\usage{
crop_image_secondary(image, new_width, new_height)
}
\description{
crop an image in R  [ for RGB or grey images ]
}
\keyword{internal}

