% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_image_rscript.R
\name{func_chech_range}
\alias{func_chech_range}
\title{function to check the range of values of an image or normalize an image}
\usage{
func_chech_range(image)
}
\description{
function to check the range of values of an image or normalize an image
}
\keyword{internal}
