% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hog_features_rscript.R
\name{func_transform}
\alias{func_transform}
\title{secondary function for HOG_apply}
\usage{
func_transform(image, folder_path, flag_type, RGB_2gray = F)
}
\description{
secondary function for HOG_apply
}
\keyword{internal}
