% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_image_rscript.R
\name{image_thresholding}
\alias{image_thresholding}
\title{image thresholding}
\usage{
image_thresholding(image, thresh)
}
\arguments{
\item{image}{matrix or 3-dimensional array}

\item{thresh}{the threshold parameter should be between 0 and 1 if the data is normalized or between 0-255 otherwise}
}
\value{
a matrix
}
\description{
image thresholding
}
\details{
This function applies thresholding to a matrix or to a 3-dimensional array.
}
\examples{

path = system.file("tmp_images", "1.png", package = "OpenImageR")

image = readImage(path)

filt = image_thresholding(image, thresh = 0.5)

}
\author{
Lampros Mouselimis
}
