% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_image_rscript.R
\name{List_2_Array}
\alias{List_2_Array}
\title{convert a list of matrices to an array of matrices}
\usage{
List_2_Array(data, verbose = FALSE)
}
\arguments{
\item{data}{a list of matrices}

\item{verbose}{if TRUE then the time taken to complete the task will be printed}
}
\value{
an array
}
\description{
convert a list of matrices to an array of matrices
}
\details{
This is a helper function mainly for the HOG and hash functions. In case that matrices are stored in a list, 
this function converts the list to an array of 2-dimensional data.
}
\examples{

lst = list(matrix(0, 100, 100), matrix(1, 100, 100))

arr = List_2_Array(lst, verbose = FALSE)

}
\author{
Lampros Mouselimis
}
