% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superpixels.R
\name{load_3d_binary}
\alias{load_3d_binary}
\title{load 3-dimensional data from a binary file}
\usage{
load_3d_binary(slic_data)
}
\arguments{
\item{slic_data}{a 3-dimensional array}
}
\description{
load 3-dimensional data from a binary file
}
\details{
This function can be used to load 3-dimensional data from a binary file. It is used in combination with the \emph{superpixels} function in case that the \emph{write_slic} parameter is not an empty string ("").
}
\examples{

\dontrun{

library(OpenImageR)

path = "/my_dir/data.bin"

res = load_3d_binary(path)

}

}
