% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_image_rscript.R
\name{norm_matrix_range}
\alias{norm_matrix_range}
\title{Normalize a matrix to specific range of values}
\usage{
norm_matrix_range(data, min_value = -1, max_value = 1)
}
\arguments{
\item{data}{a matrix}

\item{min_value}{the new minimum value for the input \emph{data}}

\item{max_value}{the new maximum value for the input \emph{data}}
}
\value{
a matrix
}
\description{
Normalize a matrix to specific range of values
}
\examples{

set.seed(1)
mt = matrix(1:48, 8, 6)

res = norm_matrix_range(mt, min_value = -1, max_value = 1)

}
