% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_image_rscript.R
\name{norm_range_gauss}
\alias{norm_range_gauss}
\title{normalize in specific range of values}
\usage{
norm_range_gauss(data, max_range = -1, min_range = 1)
}
\description{
normalize in specific range of values
}
\keyword{internal}
