% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superpixels.R
\name{RGB_to_HSV}
\alias{RGB_to_HSV}
\title{Conversion of RGB to HSV colour type}
\usage{
RGB_to_HSV(input_data)
}
\arguments{
\item{input_data}{a 3-dimensional array (RGB image) where the third dimension is equal to 3}
}
\description{
Conversion of RGB to HSV colour type
}
\details{
Meaning: RGB (Red-Green-Blue) to HSV (Hue, Saturation, Value) colour conversion
}
\examples{

library(OpenImageR)

set.seed(1)
array_3d = array(sample(1:255, 675, replace = TRUE), c(15, 15, 3))

res = RGB_to_HSV(array_3d)

}
