% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_image_rscript.R
\name{sec_gaus_bl}
\alias{sec_gaus_bl}
\title{secondary function for downsampling}
\usage{
sec_gaus_bl(image, factor, sigma, range_gauss)
}
\description{
secondary function for downsampling
}
\keyword{internal}
