% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_image_rscript.R
\name{sec_resiz_array}
\alias{sec_resiz_array}
\title{secondary function for 'resizeImage'  [ array ]}
\usage{
sec_resiz_array(image, flag = T)
}
\description{
secondary function for 'resizeImage'  [ array ]
}
\keyword{internal}

