% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash_functions.R
\name{switch_hashing}
\alias{switch_hashing}
\title{if-else function for hashing}
\usage{
switch_hashing(
  object,
  width,
  height,
  hash_size,
  highfreq_factor,
  method,
  mode,
  threads,
  resize
)
}
\description{
if-else function for hashing
}
\keyword{internal}
