% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash_functions.R
\name{switch_invariant}
\alias{switch_invariant}
\title{secondary function for invariant_hash}
\usage{
switch_invariant(method, gray_image, MODE, hash_size, highfreq_factor, resize)
}
\description{
secondary function for invariant_hash
}
\keyword{internal}
