% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_rscript.R
\name{imageShow}
\alias{imageShow}
\title{display an image}
\usage{
imageShow(file_path)
}
\arguments{
\item{file_path}{if file_path is a character string, then a shiny application is utilized. If file_path is a matrix, data.frame OR a 3-dimensional array then the grid.raster function of the base grid package is used.}
}
\value{
displays an image
}
\description{
This function displays an image
}
\details{
This function displays an image using either a character path, a 2- or a 3-dimensional object.
}
\examples{

# path = system.file("tmp_images", "1.png", package = "OpenImageR")

# imageShow(path)

}

